Move save hex button to settings menu
