+++ src/components/SettingsMenu.tsx
@@ -2,13 +2,15 @@ import {
   IconButton,
   Menu,
   MenuButton,
+  MenuDivider,
   MenuList,
   ThemeTypings,
 } from "@chakra-ui/react";
+import { useRef } from "react";
 import { RiSettings2Line } from "react-icons/ri";
 import { useIntl } from "react-intl";
 import LanguageMenuItem from "./LanguageMenuItem";
-import { useRef } from "react";
+import SaveProjectMenuItem from "./SaveProjectMenuItem";
 
 interface SettingsMenuProps {
   variant?: ThemeTypings["components"]["Menu"]["variants"];
@@ -45,6 +47,8 @@ const SettingsMenu = ({ variant = "plain", ...rest }: SettingsMenuProps) => {
           }}
         />
         <MenuList zIndex={2}>
+          <SaveProjectMenuItem />
+          <MenuDivider />
           <LanguageMenuItem finalFocusRef={settingsMenuRef} />
         </MenuList>
       </Menu>
